requirements.txt
Django==4.0.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
sentry-sdk==1.5.2
drf-yasg==1.20.0
pdfkit==1.0.0
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.0.0